Rename boundary.h/cpp to systemboundary.h/cpp
