+++ src/iomei.cpp
@@ -24,7 +24,6 @@
 #include "artic.h"
 #include "beam.h"
 #include "beatrpt.h"
-#include "boundary.h"
 #include "bracketspan.h"
 #include "breath.h"
 #include "btrem.h"
@@ -122,6 +121,7 @@
 #include "syl.h"
 #include "syllable.h"
 #include "system.h"
+#include "systemboundary.h"
 #include "tabdursym.h"
 #include "tabgrp.h"
 #include "tempo.h"
