+++ src/system.cpp
@@ -13,7 +13,6 @@
 
 //----------------------------------------------------------------------------
 
-#include "boundary.h"
 #include "comparison.h"
 #include "dir.h"
 #include "doc.h"
@@ -30,6 +29,7 @@
 #include "slur.h"
 #include "staff.h"
 #include "syl.h"
+#include "systemboundary.h"
 #include "trill.h"
 #include "verse.h"
 #include "vrv.h"
